test_main.py
from datetime import datetime

# Función para leer el valor analógico del pin
def analogRead(pin):
    # Implementación simulada para pruebas
    return 0

# Función para establecer el modo de un pin
def pinMode(pin, mode):
    # Implementación simulada para pruebas
    pass

# Función para leer la temperatura y la humedad
def readDHTSensor(port, sensor_type):
    # Implementación simulada para pruebas
    temperatura = 25.0  # Valor de ejemplo para la temperatura
    humedad = 50.0  # Valor de ejemplo para la humedad
    return temperatura, humedad

# Función para obtener la hora y la fecha actual
def obtener_hora_fecha_actual():
    now = datetime.now()
    return now.strftime("%Y-%m-%d %H:%M:%S")

# Pruebas unitarias
def test_analogRead():
    # Prueba con un pin ficticio
    pin_value = analogRead(1)
    assert pin_value == 0  # Valor simulado para la prueba

def test_pinMode():
    # Prueba con un pin ficticio y un modo ficticio
    pinMode(2, "INPUT")
    # Verifica que no se produzcan errores durante la ejecución de la función

def test_readDHTSensor():
    # Prueba con un puerto ficticio y un tipo de sensor ficticio
    temperatura, humedad = readDHTSensor(3, 0)
    assert isinstance(temperatura, float)  # Tipo de dato simulado para la prueba
    assert isinstance(humedad, float)  # Tipo de dato simulado para la prueba

def test_obtener_hora_fecha_actual():
    # Prueba la función obtener_hora_fecha_actual
    hora_fecha_actual = obtener_hora_fecha_actual()
    assert isinstance(hora_fecha_actual, str)
    assert datetime.strptime(hora_fecha_actual, "%Y-%m-%d %H:%M:%S")